ENH: Allow user-defined slicing over time or channel axes
Changes:
- Adds axes metadata parsing and exposes result in API
- Exposes optional parameter to allow slicing over channel or time axes

Addresses issue where 5D t/c/z/y/x volume subregions could not be read
with OMEZarrNGFFImageIO because Tensorstore was not provided with enough
information to map from 5D axes to 3D output image.

Could be extended in the future to allow the user to override the
default mapping of ITK dimension 0 -> "x", 1 -> "y", 2 -> "z" for
different input axes.