STYLE: Replace thread_local variables with m_TensorStoreData data member
It appears preferable that multiple OMEZarrNGFFImageIO instances do not share
their TensorStore context and store. With this commit, each OMEZarrNGFFImageIO
instance will have its own context and store, stored in its `m_TensorStoreData`
data member.

The `m_TensorStoreData` data member hides the TensorStore data from the
interface, as it is a pointer to a forward-declared struct, similar to the
"Pimpl" idiom.